Function that moves players made

+++ src/Model/Match/Point.java
@@ -1,27 +1,27 @@
 package Model.Match;
 
 public class Point {
-    private int x;
-    private int y;
+    private double x;
+    private double y;
 
     public Point(int x, int y) {
         this.x = x;
         this.y = y;
     }
 
-    public int getX() {
+    public double getX() {
         return x;
     }
 
-    public void setX(int x) {
+    public void setX(double x) {
         this.x = x;
     }
 
-    public int getY() {
+    public double getY() {
         return y;
     }
 
-    public void setY(int y) {
+    public void setY(double y) {
         this.y = y;
     }
 
@@ -29,8 +29,9 @@ public class Point {
         return Math.sqrt((p2.getX() - this.getX()) + (p2.getY() - this.getY()));
     }
 
-    public void addVector(int x, int y) {
+    public void addVector(double x, double y) {
         this.x += x;
         this.y += y;
     }
+
 }
